Modularizing growing test library. See extended notes…
I have moved rest and session tests into their own files while keeping a backup of old tests in test/index.old.js. Working on migrating those too.

Note that tests did change in this commit.
